Fixes bugs regarding protected users on twitter
Now, when a user looks at a protected user, if they
follow that user, they will be able to see their tweets.

If they don't follow that user (or if their auth tokens
have been denied), we fail gracefully.  If we are logged
in, we offer the user the chance to follow the person
they are looking at.  If they have already requested,
we show that request.  If they are not authenticated,
we tell them to log in to see protected tweets.