Significantly improves UI for posting to twitter
Adds autocomplete to @mentions that pulls in names, screennames,
and images.  Improves the layout of the post button and includes
a character counter counting down from 140.

Moves the location of inclusion of twitpost.js to only be
included if we actually are in a view where we can post.

For now, adds and points us at the jquery-ui-all package, which
includes all jquery-ui elements.  Before going live, we should
replace with a specific, trimmed-down package for faster downloads.

Adds twitpost.css to handle styling of posting box.